+++ tchannel/messages/claim.py
@@ -22,7 +22,7 @@ from __future__ import absolute_import
 
 from . import common
 from .. import rw
-from ..glossary import DEFAULT_TTL
+from ..glossary import DEFAULT_TIMEOUT
 from .base import BaseMessage
 
 
@@ -32,7 +32,7 @@ class ClaimMessage(BaseMessage):
         'tracing',
     )
 
-    def __init__(self, ttl=DEFAULT_TTL, tracing=None, id=0):
+    def __init__(self, ttl=DEFAULT_TIMEOUT, tracing=None, id=0):
         super(ClaimMessage, self).__init__(id)
         self.ttl = ttl
         self.tracing = tracing or common.Tracing(0, 0, 0, 0)